update(enum): add support for variant sized enums
